Move LIMS integrations under legacy namespace
