+++ teemi/legacy/lims/__init__.py
@@ -1,0 +1,1 @@
+"""Legacy LIMS integrations kept for backwards compatibility."""
