+++ teemi/lims/__init__.py
@@ -1,4 +1,4 @@
-"""Top-level package for teemi."""
+"""Legacy compatibility namespace for older LIMS integrations."""
 
 __author__ = """Lucas Levassor"""
 __email__ = "luclev@biosustain.dtu.dk"
